+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusSerializer.cs
@@ -26,10 +26,14 @@ internal static partial class PrometheusSerializer
     {
         if (MathHelper.IsFinite(value))
         {
+            // From https://prometheus.io/docs/specs/om/open_metrics_spec/#considerations-canonical-numbers:
+            // A warning to implementers in C and other languages that share its printf implementation:
+            // The standard precision of %f, %e and %g is only six significant digits. 17 significant
+            // digits are required for full precision, e.g. printf("%.17g", d).
 #if NET
             Span<char> span = stackalloc char[128];
 
-            var result = value.TryFormat(span, out var cchWritten, "G", CultureInfo.InvariantCulture);
+            var result = value.TryFormat(span, out var cchWritten, "G17", CultureInfo.InvariantCulture);
             Debug.Assert(result, $"{nameof(result)} should be true.");
 
             for (var i = 0; i < cchWritten; i++)
@@ -37,7 +41,7 @@ internal static partial class PrometheusSerializer
                 buffer[cursor++] = unchecked((byte)span[i]);
             }
 #else
-            cursor = WriteAsciiStringNoEscape(buffer, cursor, value.ToString(CultureInfo.InvariantCulture));
+            cursor = WriteAsciiStringNoEscape(buffer, cursor, value.ToString("G17", CultureInfo.InvariantCulture));
 #endif
         }
         else if (double.IsPositiveInfinity(value))
@@ -50,8 +54,9 @@ internal static partial class PrometheusSerializer
         }
         else
         {
+            // See https://prometheus.io/docs/instrumenting/exposition_formats/#comments-help-text-and-type-information
             Debug.Assert(double.IsNaN(value), $"{nameof(value)} should be NaN.");
-            cursor = WriteAsciiStringNoEscape(buffer, cursor, "Nan");
+            cursor = WriteAsciiStringNoEscape(buffer, cursor, "NaN");
         }
 
         return cursor;
@@ -213,7 +218,46 @@ internal static partial class PrometheusSerializer
         {
             // TODO: Attribute values should be written as their JSON representation. Extra logic may need to be added here to correctly convert other .NET types.
             // More detail: https://github.com/open-telemetry/opentelemetry-dotnet/issues/4822#issuecomment-1707328495
-            return labelValue is bool b ? b ? "true" : "false" : labelValue?.ToString() ?? string.Empty;
+            if (labelValue is bool booleanValue)
+            {
+                return booleanValue ? "true" : "false";
+            }
+            else if (labelValue is double doubleValue)
+            {
+                return DoubleToString(doubleValue);
+            }
+            else if (labelValue is float floatValue)
+            {
+                return DoubleToString(floatValue);
+            }
+
+            return labelValue?.ToString() ?? string.Empty;
+
+            static string DoubleToString(double value)
+            {
+                // From https://prometheus.io/docs/specs/om/open_metrics_spec/#considerations-canonical-numbers:
+                // A warning to implementers in C and other languages that share its printf implementation:
+                // The standard precision of %f, %e and %g is only six significant digits. 17 significant
+                // digits are required for full precision, e.g. printf("%.17g", d).
+                if (MathHelper.IsFinite(value))
+                {
+                    return value.ToString("G17", CultureInfo.InvariantCulture);
+                }
+                else if (double.IsPositiveInfinity(value))
+                {
+                    return "+Inf";
+                }
+                else if (double.IsNegativeInfinity(value))
+                {
+                    return "-Inf";
+                }
+                else
+                {
+                    // See https://prometheus.io/docs/instrumenting/exposition_formats/#comments-help-text-and-type-information
+                    Debug.Assert(double.IsNaN(value), $"{nameof(value)} should be NaN.");
+                    return "NaN";
+                }
+            }
         }
     }